+++ src/ui.c
@@ -222,6 +222,9 @@ static void create_and_attach_password_field(Config *config, UI *ui)
 {
     ui->password_input = gtk_entry_new();
     gtk_entry_set_visibility(GTK_ENTRY(ui->password_input), FALSE);
+    if (config->password_char != NULL) {
+        gtk_entry_set_invisible_char(GTK_ENTRY(ui->password_input), *config->password_char);
+    }
     gtk_entry_set_alignment(GTK_ENTRY(ui->password_input),
                             (gfloat) config->password_alignment);
     // TODO: The width is usually a little shorter than we specify. Is there a