+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## master
 
+* Add a `password-input-width` configuration option to set the number of
+  characters that should fit in the password input. This is just a suggested
+  width - GTK may render a narrower input.
 * Ensure the mouse cursor is always hidden.
 * Add a `show-image-on-all-monitors` configuration option to display the
   background image on every monitor instead of just the primary monitor.