+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## master
 
+* Add a `password-character` configuration option to set the character used to
+  mask the entered password. `-1` specifies the default character, `0` hides
+  any entered text, & the first character is used for any other value.
 * Document how to make configuration file a symlink into home directories.
 * Improve error message when configuration file is unreadable.
 * Disable the password input while authenticating. This fixes a bug causing